HBSD: Ensure signedness
Optimize the default TTL randomization calculation by making val an
unsigned integer. This way, we don't spin, trying to recalculate val if
arc4random returns a value that could be considered negative.

MFC-to:		13-STABLE
X-MFC-With:	c65edc8a6443234c9ce02b4d0d16aee6d6fae053

+++ sys/netinet/raw_ip.c
@@ -200,13 +200,13 @@ static void
 defttl_init(void)
 {
 #ifdef HBSD_RESIST_FINGERPRINTING
-	int val;
+	unsigned int val;
 
 	val = 0;
 	while (val < V_ip_minrandttl) {
-		val = (arc4random() % (254 - V_ip_minrandttl)) + 1;
+		val = (arc4random() % (254 - (unsigned)V_ip_minrandttl)) + 1;
 	}
-	V_ip_defttl = val;
+	V_ip_defttl = (int)val;
 #else
 	V_ip_defttl = IPDEFTTL;
 #endif
